requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.2
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
python-dotenv==1.0.0
qrcode==7.4.2
requests==2.31.0
SQLAlchemy==2.0.21
Werkzeug==2.3.7
python-engineio==4.7.1
python-socketio==5.9.0
email-validator==2.0.0
pyjwt==2.8.0
